requirements.txt
fastapi
uvicorn
python-dotenv
pydantic-settings
pydantic_core
SQLAlchemy
psycopg2-binary